Ensure correct lock behavior.
This commit fixes two (related) problems:

* Due to the semantics of inout stored properties, value types can't
  participate in synchronization of themselves. Therefore, change the
  ReadWriteLock protocol to require a class (disallowing structs).

* Swift doesn't guarantee stable memory addresses for stored properties,
  so instead of using raw Int32 properties for OSAtomic and friends, we
  use an explicit UnsafeMutablePointer (which is manually allocated and
  deallocated and therefore not subject to Swift moving it around behind
  the scenes).

Neither of these were problems in practice (in beta 5 or beta 6) for the
LockProtected class because it is declared as "final", and optimizations
with final classes sidestepped both of these problems. However, that
isn't guaranteed to always be the case. With these changes, the classes
should work correctly given the semantics of inout properties even in
the absence of optimizations.